Removes some handler ideas and items that fall outside the scope of the MVP - to return to later
