+++ packages/lambda/src/api/index.ts
@@ -1,5 +1,5 @@
 export * from './authorizers/index.js';
 export * from './route-resolvers/index.js';
 export * from './types/index.js';
-export * from './temp-util-functions.js';
+export * from './response-utils.js';
 export * from './http-errors.js';
