+++ packages/lambda/src/handlers/type-safe-api/handler.ts
@@ -19,7 +19,10 @@ import type {
 } from './types.js';
 import type { BaseApiRouteProps } from './types.js';
 
-import { default5xxErrorHandler } from '../../api/index.js';
+import {
+  default5xxErrorHandler,
+  makeApiGatewayResponse,
+} from '../../api/index.js';
 import { HandlerExecuteError } from '../../common/index.js';
 import {
   createRequestContext,
@@ -40,7 +43,7 @@ export type RouteResolver = (
 interface TypeSafeApiHandlerReturn extends ICreatableReturn {
   (
     routes: TypedApiRouteConfig,
-    hooks?: TypeSafeApiHandlerHooks
+    hooks?: Partial<TypeSafeApiHandlerHooks>
   ): LambdaEntryPoint<APIGatewayProxyEventV2, APIGatewayProxyResultV2 | void>;
 }
 
@@ -52,12 +55,6 @@ type PrivateResponseContext = ResponseContext & {
 export interface TypeSafeApiHandlerConfig extends ICreatableConfig {
   devMode?: true;
   name?: `${string}Handler`;
-
-  // Other ideas
-  // basePath?: string;
-  // msTimeoutBudget?: number;
-  // openApiRoute?: true;
-  // respectMethodOverrideHeader?: true;
 }
 
 export const TypeSafeApiHandler = (
@@ -164,79 +161,3 @@ export const TypeSafeApiHandler = (
     }
   };
 };
-
-/**
- * Builds an APIGatewayProxyResultV2 from the result object and response context
- * object.
- * @param result the result object from the route handler function
- * @param responseContext the response context object that originates from the
- * handler
- * @returns an APIGatewayProxyResultV2
- */
-function makeApiGatewayResponse(
-  result: ApiHandlerReturn,
-  responseContext: PrivateResponseContext
-): APIGatewayProxyResultV2 {
-  const response: APIGatewayProxyResultV2 = {
-    body: JSON.stringify(result.body),
-    headers: {
-      ...responseContext.headers,
-      ...result.headers,
-    },
-    statusCode: responseContext._statusCode,
-  };
-
-  response.cookies = buildCookiesFromObject({
-    ...responseContext.cookies,
-    ...result.cookies,
-  });
-
-  return response;
-}
-
-// TODO: enhancement: expose handler option for encoding cookies
-/**
- * Builds an array of cookie strings from an object of cookies options.
- * @param cookieObject an object of cookie options
- * @returns an array of cookie strings
- */
-function buildCookiesFromObject(
-  cookieObject: ApiHandlerReturn['cookies']
-): string[] | undefined {
-  if (!cookieObject) return undefined;
-  if (Object.keys(cookieObject).length === 0) return undefined;
-
-  let cookieString = '';
-  const cookiesArray = [];
-  for (const cookie in cookieObject) {
-    // eslint-disable-next-line security/detect-object-injection
-    cookieString = `${cookie}=${cookieObject[cookie].value}`;
-
-    // eslint-disable-next-line security/detect-object-injection
-    if (cookieObject[cookie].options) {
-      const {
-        domain,
-        expiresUnix,
-        httpOnly,
-        maxAge,
-        path,
-        sameSite,
-        secure,
-        // eslint-disable-next-line security/detect-object-injection
-      } = cookieObject[cookie].options || {};
-
-      if (domain) cookieString += `; Domain=${domain}`;
-      if (expiresUnix)
-        cookieString += `; Expires=${new Date(expiresUnix).toUTCString()}`;
-      if (httpOnly) cookieString += `; HttpOnly`;
-      if (maxAge) cookieString += `; Max-Age=${maxAge}`;
-      if (path) cookieString += `; Path=${path}`;
-      if (sameSite) cookieString += `; SameSite=${sameSite}`;
-      if (secure) cookieString += `; Secure`;
-    }
-
-    cookiesArray.push(cookieString);
-  }
-
-  return cookiesArray;
-}
